requirements.txt
django
plotly
pandas
numpy
django-plotly-dash
channels 
daphne
redis
django-redis
channels-redis
psycopg2
dash
djangorestframework